requirements.txt
pytest-playwright
google-generativeai
ollama